requirements.txt
flask
GitPython
langchain-openai
langchain-core